230112 - ↓작업 내용↓ 1. ComponentUI 설계
2. 각 컴포넌트 별 ComponentUI 추가

3. MeshRenderUI 작업 중

4. Modal 방식 UI 기능 추가

+++ Project/Engine/CTexture.h
@@ -21,8 +21,8 @@ private:
 
 
 public:
-    float Width() { return m_Desc.Width; }
-    float Height() { return m_Desc.Height; }
+    float Width() { return (float)m_Desc.Width; }
+    float Height() { return (float)m_Desc.Height; }
 
 private:
     virtual int Load(const wstring& _strFilePath) override;
